chore: move usecase port into port database
Story: [ch18]
