--- packages/server/accounts/src/domain/use-cases/ports/errors/index.ts
@@ -1,1 +1,0 @@
-export * from '@/domain/use-cases/ports/errors/infra';
