+++ packages/server/accounts/src/ports/database/fakes/index.ts
@@ -1,0 +1,1 @@
+export * from '@/ports/database/fakes/users-repo';
