+++ packages/server/accounts/src/ports/database/index.ts
@@ -1,0 +1,1 @@
+export * from '@/ports/database/users-repo';
