+++ packages/server/accounts/src/ports/errors/index.ts
@@ -1,0 +1,1 @@
+export * from '@/ports/errors/infra';
